+++ src/praisonai/praisonai/integrations/compute/e2b.py
@@ -1,228 +1,18 @@
+"""Moved to :mod:`praisonai_sandbox.compute.e2b`.
+
+The vendor implementations moved into praisonai-sandbox, which already carries
+one optional extra per vendor -- reaching one provider no longer requires the
+whole wrapper.
+
+This module *becomes* the new one rather than re-exporting from it. A plain
+``from ... import X`` shim looks equivalent until someone monkeypatches a
+module-level name through the old path: the patch lands on the shim and the
+real module never sees it. Aliasing in ``sys.modules`` keeps every existing
+import, attribute and patch working against one shared module object.
 """
-E2B Compute Provider — cloud sandbox-based compute for managed agents.
 
-Uses E2B SDK for secure, isolated sandbox environments with <1s boot.
+import sys
 
-Requires: ``pip install e2b``
-Environment: ``E2B_API_KEY``
-"""
+from praisonai_sandbox.compute import e2b as _moved
 
-from ._sync_base import SyncComputeProvider
-import logging
-import os
-import time
-import uuid
-from typing import Any, Dict, List
-
-logger = logging.getLogger(__name__)
-
-
-class E2BCompute(SyncComputeProvider):
-    """E2B cloud sandbox compute provider.
-
-    Satisfies ``ComputeProviderProtocol`` (Core SDK).
-    Uses E2B SDK for secure, isolated sandbox environments.
-
-    Example::
-
-        from praisonaiagents.managed import ComputeConfig
-        from praisonai.integrations.compute.e2b import E2BCompute
-
-        compute = E2BCompute()
-        config = ComputeConfig(
-            image="python:3.12-slim",
-            packages={"pip": ["pandas"]},
-        )
-        info = await compute.provision(config)
-        result = await compute.execute(info.instance_id, "python -c 'print(1+1)'")
-        await compute.shutdown(info.instance_id)
-    """
-
-    def __init__(
-        self,
-        api_key: str = "",
-    ) -> None:
-        self._api_key = api_key or os.environ.get("E2B_API_KEY", "")
-        self._sandboxes: Dict[str, Dict[str, Any]] = {}
-
-    @property
-    def provider_name(self) -> str:
-        return "e2b"
-
-    @property
-    def is_available(self) -> bool:
-        return bool(self._api_key)
-
-
-    def _provision_sync(self, config) -> Any:
-        from praisonaiagents.managed.protocols import InstanceInfo, InstanceStatus
-
-        try:
-            from e2b import Sandbox
-        except ImportError:
-            raise ImportError("E2B SDK required. Install with: pip install e2b")
-
-        instance_id = f"e2b_{uuid.uuid4().hex[:12]}"
-
-        # E2B uses named templates, not Docker image refs. An explicit
-        # `metadata["e2b_template"]` wins; failing that, a caller who set
-        # `image=` to something other than the provider-agnostic default is
-        # asking for a specific environment, so honour it as the template
-        # rather than silently booting E2B's default.
-        from praisonaiagents.managed.protocols import ComputeConfig as _CC
-
-        template = config.metadata.get("e2b_template") if config.metadata else None
-        if not template and config.image and config.image != _CC.image:
-            template = config.image
-
-        sandbox = Sandbox.create(
-            template=template,
-            timeout=max(config.idle_timeout_s, 60),
-            envs=config.env or None,
-            api_key=self._api_key,
-        )
-
-        self._sandboxes[instance_id] = {
-            "sandbox": sandbox,
-            "sandbox_id": sandbox.sandbox_id,
-            "config": config,
-            "created_at": time.time(),
-        }
-
-        # Install packages
-        if config.packages:
-            self._install_packages_sync(sandbox, config.packages)
-
-        logger.info("[e2b_compute] provisioned: %s sandbox=%s", instance_id, sandbox.sandbox_id)
-
-        return InstanceInfo(
-            instance_id=instance_id,
-            status=InstanceStatus.RUNNING,
-            endpoint=f"e2b://{sandbox.sandbox_id}",
-            provider="e2b",
-            created_at=time.time(),
-        )
-
-    async def shutdown(self, instance_id: str) -> None:
-        import asyncio
-        loop = asyncio.get_running_loop()
-        await loop.run_in_executor(None, self._shutdown_sync, instance_id)
-
-    def _shutdown_sync(self, instance_id: str) -> None:
-        info = self._sandboxes.pop(instance_id, None)
-        if info:
-            try:
-                info["sandbox"].kill()
-            except Exception as e:
-                logger.warning("[e2b_compute] shutdown error: %s", e)
-            logger.info("[e2b_compute] shutdown: %s", instance_id)
-
-    async def get_status(self, instance_id: str) -> Any:
-        from praisonaiagents.managed.protocols import InstanceInfo, InstanceStatus
-
-        info = self._sandboxes.get(instance_id)
-        if not info:
-            return InstanceInfo(
-                instance_id=instance_id,
-                status=InstanceStatus.STOPPED,
-                provider="e2b",
-            )
-        try:
-            running = info["sandbox"].is_running()
-        except Exception:
-            running = False
-
-        return InstanceInfo(
-            instance_id=instance_id,
-            status=InstanceStatus.RUNNING if running else InstanceStatus.STOPPED,
-            endpoint=f"e2b://{info['sandbox_id']}",
-            provider="e2b",
-            created_at=info.get("created_at", 0),
-        )
-
-
-    def _execute_sync(
-        self, instance_id: str, command: str, timeout: int,
-    ) -> Dict[str, Any]:
-        info = self._sandboxes.get(instance_id)
-        if not info:
-            return {"stdout": "", "stderr": "Instance not found", "exit_code": -1}
-
-        sandbox = info["sandbox"]
-        try:
-            result = sandbox.commands.run(command, timeout=timeout)
-            return {
-                "stdout": result.stdout or "",
-                "stderr": result.stderr or "",
-                "exit_code": result.exit_code,
-            }
-        except Exception as e:
-            return {"stdout": "", "stderr": str(e), "exit_code": -1}
-
-
-    def _upload_sync(self, instance_id: str, local_path: str, remote_path: str) -> bool:
-        info = self._sandboxes.get(instance_id)
-        if not info:
-            return False
-        try:
-            with open(local_path, "rb") as f:
-                content = f.read()
-            info["sandbox"].files.write(remote_path, content)
-            return True
-        except Exception as e:
-            logger.error("[e2b_compute] upload failed: %s", e)
-            return False
-
-
-    def _download_sync(self, instance_id: str, remote_path: str, local_path: str) -> bool:
-        info = self._sandboxes.get(instance_id)
-        if not info:
-            return False
-        try:
-            content = info["sandbox"].files.read(remote_path)
-            with open(local_path, "w") as f:
-                f.write(content)
-            return True
-        except Exception as e:
-            logger.error("[e2b_compute] download failed: %s", e)
-            return False
-
-    async def list_instances(self) -> List[Any]:
-        from praisonaiagents.managed.protocols import InstanceInfo, InstanceStatus
-
-        result = []
-        for iid, info in self._sandboxes.items():
-            try:
-                running = info["sandbox"].is_running()
-            except Exception:
-                running = False
-            if running:
-                result.append(InstanceInfo(
-                    instance_id=iid,
-                    status=InstanceStatus.RUNNING,
-                    endpoint=f"e2b://{info['sandbox_id']}",
-                    provider="e2b",
-                    created_at=info.get("created_at", 0),
-                ))
-        return result
-
-    def _install_packages_sync(self, sandbox, packages: Dict[str, list]) -> None:
-        pip_pkgs = packages.get("pip", [])
-        if pip_pkgs:
-            cmd = f"pip install -q {' '.join(pip_pkgs)}"
-            logger.info("[e2b_compute] installing pip: %s", pip_pkgs)
-            try:
-                result = sandbox.commands.run(cmd, timeout=120)
-                if result.exit_code != 0:
-                    logger.warning("[e2b_compute] pip install failed: %s", result.stderr)
-            except Exception as e:
-                logger.warning("[e2b_compute] pip install error: %s", e)
-
-        npm_pkgs = packages.get("npm", [])
-        if npm_pkgs:
-            cmd = f"npm install -g {' '.join(npm_pkgs)}"
-            logger.info("[e2b_compute] installing npm: %s", npm_pkgs)
-            try:
-                sandbox.commands.run(cmd, timeout=120)
-            except Exception as e:
-                logger.warning("[e2b_compute] npm install error: %s", e)
+sys.modules[__name__] = _moved
